+++ bundle-polkadot-util-crypto.js
@@ -6,12 +6,12 @@
 
     const global = typeof globalThis !== "undefined" ? globalThis : typeof self !== "undefined" ? self : window;
 
-    const packageInfo$2 = { name: '@polkadot/x-global', path: (({ url: (typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href)) }) && (typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))) ? new URL((typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))).pathname.substring(0, new URL((typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))).pathname.lastIndexOf('/') + 1) : 'auto', type: 'esm', version: '11.0.2' };
+    const packageInfo$2 = { name: '@polkadot/x-global', path: (({ url: (typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href)) }) && (typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))) ? new URL((typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))).pathname.substring(0, new URL((typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))).pathname.lastIndexOf('/') + 1) : 'auto', type: 'esm', version: '11.1.1' };
 
     function evaluateThis(fn) {
         return fn('return this');
     }
-    const xglobal = (typeof globalThis !== 'undefined'
+    const xglobal =  (typeof globalThis !== 'undefined'
         ? globalThis
         : typeof global !== 'undefined'
             ? global
@@ -1675,8 +1675,6 @@
         return (kind === "a" ? f.call(receiver, value) : f ? f.value = value : state.set(receiver, value)), value;
     }
 
-    var commonjsGlobal = typeof globalThis !== 'undefined' ? globalThis : typeof window !== 'undefined' ? window : typeof global !== 'undefined' ? global : typeof self !== 'undefined' ? self : {};
-
     function getDefaultExportFromCjs (x) {
     	return x && x.__esModule && Object.prototype.hasOwnProperty.call(x, 'default') ? x['default'] : x;
     }
@@ -1717,7 +1715,7 @@
 
     Object.defineProperty(packageInfo$1, "__esModule", { value: true });
     packageInfo$1.packageInfo = void 0;
-    packageInfo$1.packageInfo = { name: '@polkadot/x-randomvalues', path: typeof __dirname === 'string' ? __dirname : 'auto', type: 'cjs', version: '11.0.2' };
+    packageInfo$1.packageInfo = { name: '@polkadot/x-randomvalues', path: typeof __dirname === 'string' ? __dirname : 'auto', type: 'cjs', version: '11.1.1' };
 
     (function (exports) {
     	Object.defineProperty(exports, "__esModule", { value: true });
@@ -2377,7 +2375,7 @@
     cryptoWaitReady().catch(() => {
     });
 
-    const packageInfo = { name: '@polkadot/util-crypto', path: (({ url: (typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href)) }) && (typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))) ? new URL((typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))).pathname.substring(0, new URL((typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))).pathname.lastIndexOf('/') + 1) : 'auto', type: 'esm', version: '11.0.2' };
+    const packageInfo = { name: '@polkadot/util-crypto', path: (({ url: (typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href)) }) && (typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))) ? new URL((typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))).pathname.substring(0, new URL((typeof document === 'undefined' && typeof location === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-util-crypto.js', document.baseURI).href))).pathname.lastIndexOf('/') + 1) : 'auto', type: 'esm', version: '11.1.1' };
 
     /*! scure-base - MIT License (c) 2022 Paul Miller (paulmillr.com) */
     function assertNumber(n) {
@@ -2778,7 +2776,7 @@
             try {
                 return validate(value, ipfsCompat);
             }
-            catch (error) {
+            catch {
                 return false;
             }
         };
@@ -2839,13 +2837,13 @@
             assert.number(outputLen);
             assert.number(keyLen);
             if (outputLen < 0 || outputLen > keyLen)
-                throw new Error('Blake2: outputLen bigger than keyLen');
+                throw new Error('outputLen bigger than keyLen');
             if (opts.key !== undefined && (opts.key.length < 1 || opts.key.length > keyLen))
-                throw new Error(`Key should be up 1..${keyLen} byte long or undefined`);
+                throw new Error(`key must be up 1..${keyLen} byte long or undefined`);
             if (opts.salt !== undefined && opts.salt.length !== saltLen)
-                throw new Error(`Salt should be ${saltLen} byte long or undefined`);
+                throw new Error(`salt must be ${saltLen} byte long or undefined`);
             if (opts.personalization !== undefined && opts.personalization.length !== persLen)
-                throw new Error(`Personalization should be ${persLen} byte long or undefined`);
+                throw new Error(`personalization must be ${persLen} byte long or undefined`);
             this.buffer32 = u32$1((this.buffer = new Uint8Array(blockLen)));
         }
         update(data) {
@@ -4856,6 +4854,9 @@
         origintrail: [
             '0xe7e0962324a3b86c83404dbea483f25fb5dab4c224791c81b756cfc948006174'
         ],
+        p3d: [
+            '0x6c5894837ad89b6d92b114a2fb3eafa8fe3d26a54848e3447015442cd6ef4e66'
+        ],
         parallel: [
             '0xe61a41c53f5dcd0beb09df93b34402aada44cb05117b71059cce40a2723a4e97'
         ],
@@ -5236,11 +5237,23 @@
 
     const keyHdkdEcdsa = createSeedDeriveFn(secp256k1PairFromSeed, secp256k1DeriveHard);
 
-    var ed2curveExports = {};
-    var ed2curve$1 = {
-      get exports(){ return ed2curveExports; },
-      set exports(v){ ed2curveExports = v; },
-    };
+    const HDKD = util.compactAddLength(util.stringToU8a('Ed25519HDKD'));
+    function ed25519DeriveHard(seed, chainCode) {
+        if (!util.isU8a(chainCode) || chainCode.length !== 32) {
+            throw new Error('Invalid chainCode passed to derive');
+        }
+        return blake2AsU8a(util.u8aConcat(HDKD, seed, chainCode));
+    }
+
+    function randomAsU8a(length = 32) {
+        return browser.getRandomValues(new Uint8Array(length));
+    }
+    const randomAsHex = createAsHex(randomAsU8a);
+
+    const BN_53 = new util.BN(0b11111111111111111111111111111111111111111111111111111);
+    function randomAsNumber() {
+        return util.hexToBn(randomAsHex(8)).and(BN_53).toNumber();
+    }
 
     function commonjsRequire(path) {
     	throw new Error('Could not dynamically require "' + path + '". Please configure the dynamicRequireTargets or/and ignoreDynamicRequires option of @rollup/plugin-commonjs appropriately for this require call to work.');
@@ -7314,231 +7327,6 @@
     } (naclFast));
     const nacl = naclFastExports;
 
-    (function (module) {
-    	(function(root, f) {
-    	  if (module.exports) module.exports = f(naclFastExports);
-    	  else root.ed2curve = f(root.nacl);
-    	}(commonjsGlobal, function(nacl) {
-    	  if (!nacl) throw new Error('tweetnacl not loaded');
-    	  var gf = function(init) {
-    	    var i, r = new Float64Array(16);
-    	    if (init) for (i = 0; i < init.length; i++) r[i] = init[i];
-    	    return r;
-    	  };
-    	  var gf0 = gf(),
-    	      gf1 = gf([1]),
-    	      D = gf([0x78a3, 0x1359, 0x4dca, 0x75eb, 0xd8ab, 0x4141, 0x0a4d, 0x0070, 0xe898, 0x7779, 0x4079, 0x8cc7, 0xfe73, 0x2b6f, 0x6cee, 0x5203]),
-    	      I = gf([0xa0b0, 0x4a0e, 0x1b27, 0xc4ee, 0xe478, 0xad2f, 0x1806, 0x2f43, 0xd7a7, 0x3dfb, 0x0099, 0x2b4d, 0xdf0b, 0x4fc1, 0x2480, 0x2b83]);
-    	  function car25519(o) {
-    	    var c;
-    	    var i;
-    	    for (i = 0; i < 16; i++) {
-    	      o[i] += 65536;
-    	      c = Math.floor(o[i] / 65536);
-    	      o[(i+1)*(i<15?1:0)] += c - 1 + 37 * (c-1) * (i===15?1:0);
-    	      o[i] -= (c * 65536);
-    	    }
-    	  }
-    	  function sel25519(p, q, b) {
-    	    var t, c = ~(b-1);
-    	    for (var i = 0; i < 16; i++) {
-    	      t = c & (p[i] ^ q[i]);
-    	      p[i] ^= t;
-    	      q[i] ^= t;
-    	    }
-    	  }
-    	  function unpack25519(o, n) {
-    	    var i;
-    	    for (i = 0; i < 16; i++) o[i] = n[2*i] + (n[2*i+1] << 8);
-    	    o[15] &= 0x7fff;
-    	  }
-    	  function A(o, a, b) {
-    	    var i;
-    	    for (i = 0; i < 16; i++) o[i] = (a[i] + b[i])|0;
-    	  }
-    	  function Z(o, a, b) {
-    	    var i;
-    	    for (i = 0; i < 16; i++) o[i] = (a[i] - b[i])|0;
-    	  }
-    	  function M(o, a, b) {
-    	    var i, j, t = new Float64Array(31);
-    	    for (i = 0; i < 31; i++) t[i] = 0;
-    	    for (i = 0; i < 16; i++) {
-    	      for (j = 0; j < 16; j++) {
-    	        t[i+j] += a[i] * b[j];
-    	      }
-    	    }
-    	    for (i = 0; i < 15; i++) {
-    	      t[i] += 38 * t[i+16];
-    	    }
-    	    for (i = 0; i < 16; i++) o[i] = t[i];
-    	    car25519(o);
-    	    car25519(o);
-    	  }
-    	  function S(o, a) {
-    	    M(o, a, a);
-    	  }
-    	  function inv25519(o, i) {
-    	    var c = gf();
-    	    var a;
-    	    for (a = 0; a < 16; a++) c[a] = i[a];
-    	    for (a = 253; a >= 0; a--) {
-    	      S(c, c);
-    	      if(a !== 2 && a !== 4) M(c, c, i);
-    	    }
-    	    for (a = 0; a < 16; a++) o[a] = c[a];
-    	  }
-    	  function pack25519(o, n) {
-    	    var i, j, b;
-    	    var m = gf(), t = gf();
-    	    for (i = 0; i < 16; i++) t[i] = n[i];
-    	    car25519(t);
-    	    car25519(t);
-    	    car25519(t);
-    	    for (j = 0; j < 2; j++) {
-    	      m[0] = t[0] - 0xffed;
-    	      for (i = 1; i < 15; i++) {
-    	        m[i] = t[i] - 0xffff - ((m[i-1]>>16) & 1);
-    	        m[i-1] &= 0xffff;
-    	      }
-    	      m[15] = t[15] - 0x7fff - ((m[14]>>16) & 1);
-    	      b = (m[15]>>16) & 1;
-    	      m[14] &= 0xffff;
-    	      sel25519(t, m, 1-b);
-    	    }
-    	    for (i = 0; i < 16; i++) {
-    	      o[2*i] = t[i] & 0xff;
-    	      o[2*i+1] = t[i] >> 8;
-    	    }
-    	  }
-    	  function par25519(a) {
-    	    var d = new Uint8Array(32);
-    	    pack25519(d, a);
-    	    return d[0] & 1;
-    	  }
-    	  function vn(x, xi, y, yi, n) {
-    	    var i, d = 0;
-    	    for (i = 0; i < n; i++) d |= x[xi + i] ^ y[yi + i];
-    	    return (1 & ((d - 1) >>> 8)) - 1;
-    	  }
-    	  function crypto_verify_32(x, xi, y, yi) {
-    	    return vn(x, xi, y, yi, 32);
-    	  }
-    	  function neq25519(a, b) {
-    	    var c = new Uint8Array(32), d = new Uint8Array(32);
-    	    pack25519(c, a);
-    	    pack25519(d, b);
-    	    return crypto_verify_32(c, 0, d, 0);
-    	  }
-    	  function pow2523(o, i) {
-    	    var c = gf();
-    	    var a;
-    	    for (a = 0; a < 16; a++) c[a] = i[a];
-    	    for (a = 250; a >= 0; a--) {
-    	      S(c, c);
-    	      if (a !== 1) M(c, c, i);
-    	    }
-    	    for (a = 0; a < 16; a++) o[a] = c[a];
-    	  }
-    	  function set25519(r, a) {
-    	    var i;
-    	    for (i = 0; i < 16; i++) r[i] = a[i] | 0;
-    	  }
-    	  function unpackneg(r, p) {
-    	    var t = gf(), chk = gf(), num = gf(),
-    	      den = gf(), den2 = gf(), den4 = gf(),
-    	      den6 = gf();
-    	    set25519(r[2], gf1);
-    	    unpack25519(r[1], p);
-    	    S(num, r[1]);
-    	    M(den, num, D);
-    	    Z(num, num, r[2]);
-    	    A(den, r[2], den);
-    	    S(den2, den);
-    	    S(den4, den2);
-    	    M(den6, den4, den2);
-    	    M(t, den6, num);
-    	    M(t, t, den);
-    	    pow2523(t, t);
-    	    M(t, t, num);
-    	    M(t, t, den);
-    	    M(t, t, den);
-    	    M(r[0], t, den);
-    	    S(chk, r[0]);
-    	    M(chk, chk, den);
-    	    if (neq25519(chk, num)) M(r[0], r[0], I);
-    	    S(chk, r[0]);
-    	    M(chk, chk, den);
-    	    if (neq25519(chk, num)) return -1;
-    	    if (par25519(r[0]) === (p[31] >> 7)) Z(r[0], gf0, r[0]);
-    	    M(r[3], r[0], r[1]);
-    	    return 0;
-    	  }
-    	  function convertPublicKey(pk) {
-    	    var z = new Uint8Array(32),
-    	      q = [gf(), gf(), gf(), gf()],
-    	      a = gf(), b = gf();
-    	    if (unpackneg(q, pk)) return null;
-    	    var y = q[1];
-    	    A(a, gf1, y);
-    	    Z(b, gf1, y);
-    	    inv25519(b, b);
-    	    M(a, a, b);
-    	    pack25519(z, a);
-    	    return z;
-    	  }
-    	  function convertSecretKey(sk) {
-    	    var d = new Uint8Array(64), o = new Uint8Array(32), i;
-    	    nacl.lowlevel.crypto_hash(d, sk, 32);
-    	    d[0] &= 248;
-    	    d[31] &= 127;
-    	    d[31] |= 64;
-    	    for (i = 0; i < 32; i++) o[i] = d[i];
-    	    for (i = 0; i < 64; i++) d[i] = 0;
-    	    return o;
-    	  }
-    	  function convertKeyPair(edKeyPair) {
-    	    var publicKey = convertPublicKey(edKeyPair.publicKey);
-    	    if (!publicKey) return null;
-    	    return {
-    	      publicKey: publicKey,
-    	      secretKey: convertSecretKey(edKeyPair.secretKey)
-    	    };
-    	  }
-    	  return {
-    	    convertPublicKey: convertPublicKey,
-    	    convertSecretKey: convertSecretKey,
-    	    convertKeyPair: convertKeyPair,
-    	  };
-    	}));
-    } (ed2curve$1));
-    const ed2curve = ed2curveExports;
-
-    function convertSecretKeyToCurve25519(secretKey) {
-        return ed2curve.convertSecretKey(secretKey);
-    }
-    function convertPublicKeyToCurve25519(publicKey) {
-        return util.assertReturn(ed2curve.convertPublicKey(publicKey), 'Unable to convert publicKey to ed25519');
-    }
-
-    const HDKD = util.compactAddLength(util.stringToU8a('Ed25519HDKD'));
-    function ed25519DeriveHard(seed, chainCode) {
-        if (!util.isU8a(chainCode) || chainCode.length !== 32) {
-            throw new Error('Invalid chainCode passed to derive');
-        }
-        return blake2AsU8a(util.u8aConcat(HDKD, seed, chainCode));
-    }
-
-    function randomAsU8a(length = 32) {
-        return browser.getRandomValues(new Uint8Array(length));
-    }
-    const randomAsHex = createAsHex(randomAsU8a);
-
-    const BN_53 = new util.BN(0b11111111111111111111111111111111111111111111111111111);
-    function randomAsNumber() {
-        return util.hexToBn(randomAsHex(8)).and(BN_53).toNumber();
-    }
-
     function ed25519PairFromSeed(seed, onlyJs) {
         if (!onlyJs && isReady()) {
             const full = ed25519KeypairFromSeed(seed);
@@ -7964,7 +7752,7 @@
         try {
             return validateAddress(address, ignoreChecksum, ss58Format);
         }
-        catch (error) {
+        catch {
             return false;
         }
     }
@@ -8209,7 +7997,7 @@
             const I = hmacShaAsU8a(hd.chainCode, data, 512);
             return createCoded(secp256k1PrivateKeyTweakAdd(hd.secretKey, I.slice(0, 32)), I.slice(32));
         }
-        catch (err) {
+        catch {
             return deriveChild(hd, index + 1);
         }
     }
@@ -8357,7 +8145,7 @@
         try {
             mnemonicToEntropy$1(mnemonic);
         }
-        catch (e) {
+        catch {
             return false;
         }
         return true;
@@ -8753,7 +8541,7 @@
                     return true;
                 }
             }
-            catch (error) {
+            catch {
             }
             return false;
         });
@@ -8775,7 +8563,7 @@
                 sr25519: () => sr25519Verify(message, signature.subarray(1), publicKey)
             }[type]();
         }
-        catch (error) {
+        catch {
         }
         return result;
     }
@@ -8941,8 +8729,6 @@
     exports.blake2AsU8a = blake2AsU8a;
     exports.checkAddress = checkAddress;
     exports.checkAddressChecksum = checkAddressChecksum;
-    exports.convertPublicKeyToCurve25519 = convertPublicKeyToCurve25519;
-    exports.convertSecretKeyToCurve25519 = convertSecretKeyToCurve25519;
     exports.createKeyDerived = createKeyDerived;
     exports.createKeyMulti = createKeyMulti;
     exports.cryptoIsReady = cryptoIsReady;